summary/pyproject.toml
[project]
name = "summary"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.28.0",
    "kokkai-db",
    "psycopg[binary]>=3.2.9",
    "python-dotenv>=1.1.1",
    "pytimeparse>=1.1.8",
    "sqlalchemy>=2.0.42",
    "tenacity>=8.5.0",
]

[tool.uv.sources]
kokkai-db = { path = "../db" }

[tool.ruff]
extend = '../ruff_config.toml'

[tool.pyright]
venvPath = "."
venv = ".venv"